Refactor: 특정 날짜의 몸무게 타입을 Double타입으로 변경

+++ src/main/java/com/examle/sikmogilbackend/record/Calendar/api/dto/FindCalendarByDateDTO.java
@@ -16,7 +16,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class FindCalendarByDateDTO {
     private String diaryDate;
-    private Long diaryWeight;
+    private Double diaryWeight;
     private String diaryText;
     private Long canEatCalorie;
     private Long waterIntake;
